backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7

Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1

bcrypt==4.2.1
